Move CallableValues to prevent duplicate code
